requirements.txt
langchain==0.1.12
langchain-community==0.0.28
pandas
Flask==2.3.2
Flask-RESTful==0.3.10
sentence-transformers==2.6.0
pymongo==4.6.3
chromadb==0.4.24
flask_cors==4.0.0
gunicorn
celery
pika
